+++ src/app/playlists/components/DiggingPli.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 
 import pli from "../types/pli";
 
-import DigCount from "./Digcount_temp";
+import DigCount from "./Digcount";
 
 import { playArt } from "@/assets/images";
