+++ src/app/playlists/components/Frame80.tsx
@@ -3,7 +3,7 @@
 import Image from "next/image";
 import { readableColor } from "polished";
 
-import DigCount from "./Digcount_temp";
+import DigCount from "./Digcount";
 
 import { playArt } from "@/assets/images";
